game resize tests
